initial

+++ index.ts
@@ -4,7 +4,7 @@ import fs from 'fs';
 import bt from './bot';
 import Hlp from './helpers';
 import config from './config'
-
+let objj: any = {};
 // 2 global dependencies
 import { Scenes, session, Telegraf } from "telegraf";
 
@@ -18,7 +18,7 @@ import jvStarter from './starters/jvstarter'
 import goStarter from './starters/gostarter'
 
 // this will run web server and always make it alive
-keep_alive()
+keep_alive(objj)
 
 // env variable 
 require('dotenv').config()
@@ -33,91 +33,107 @@ const { enter, leave } = Scenes.Stage;
 let func: any = {};
 
 // Helper function which replace bot username if exists in command
-function cmdd(ctx:any){
-    ctx.message.text = ctx.message.text.replace(new RegExp("^\\"+ config.startSymbol + "[a-zA-Z0-9]{2,9}@" + ctx.botInfo.username, 'i'), 
-  (match: any) => match.replace("@" + ctx.botInfo.username, ""))
+function cmdd(ctx: any) {
+  ctx.message.text = ctx.message.text.replace(new RegExp("^\\" + config.startSymbol + "[a-zA-Z0-9]{2,9}@" + ctx.botInfo.username, 'i'),
+    (match: any) => match.replace("@" + ctx.botInfo.username, ""))
 }
 
 // All scenes 
 let pyScene = new Scenes.BaseScene<Scenes.SceneContext>("py");
-pyScene.enter(async (ctx: any) => { 
+pyScene.enter(async (ctx: any) => {
   cmdd(ctx);
-  if(await startcheck(ctx, 'py python')) 
+  if (await startcheck(ctx, 'py python'))
     return;
-  await pyStarter(bot, ctx) });
+  await pyStarter(bot, ctx)
+});
 
-pyScene.on("message", async (ctx: any) => { 
+pyScene.on("message", async (ctx: any) => {
   cmdd(ctx);
-  if(await startcheck(ctx, 'py python')) return;
-  await pyStarter(bot, ctx) });
+  if (await startcheck(ctx, 'py python')) return;
+  await pyStarter(bot, ctx)
+});
 
 let cppScene = new Scenes.BaseScene<Scenes.SceneContext>("cpp");
 
-cppScene.enter(async (ctx: any) => { 
+cppScene.enter(async (ctx: any) => {
   cmdd(ctx)
-  if(await startcheck(ctx, 'cpp cplus')) return;
-    
-  await cppStarter(bot, ctx) });
+  if (await startcheck(ctx, 'cpp cplus')) return;
 
-cppScene.on("message", async (ctx: any) => { 
-   cmdd(ctx)
-    if(await startcheck(ctx, 'cpp cplus')) return;
-  await cppStarter(bot, ctx) });
+  await cppStarter(bot, ctx)
+});
+
+cppScene.on("message", async (ctx: any) => {
+  cmdd(ctx)
+  if (await startcheck(ctx, 'cpp cplus')) return;
+  await cppStarter(bot, ctx)
+});
 
 let cScene = new Scenes.BaseScene<Scenes.SceneContext>("code");
-cScene.enter(async (ctx: any) => { 
-   cmdd(ctx)
-    if(await startcheck(ctx, 'cc code')) return;
-  await cStarter(bot, ctx); });
+cScene.enter(async (ctx: any) => {
+  cmdd(ctx)
+  if (await startcheck(ctx, 'cc code')) return;
+  await cStarter(bot, ctx);
+});
 
-cScene.on("message", async (ctx: any) => { 
-   cmdd(ctx)
-     if(await startcheck(ctx, 'cc code')) return;
-  await cStarter(bot, ctx) });
+cScene.on("message", async (ctx: any) => {
+  cmdd(ctx)
+  if (await startcheck(ctx, 'cc code')) return;
+  await cStarter(bot, ctx)
+});
 
 let jvScene = new Scenes.BaseScene<Scenes.SceneContext>("jv");
-jvScene.enter(async (ctx: any) => { 
-   cmdd(ctx)
-     if(await startcheck(ctx, 'jv java')) return;
-  await jvStarter(bot, ctx) });
+jvScene.enter(async (ctx: any) => {
+  cmdd(ctx)
+  if (await startcheck(ctx, 'jv java')) return;
+  await jvStarter(bot, ctx)
+});
 
-jvScene.on("message", async (ctx: any) => { 
-   cmdd(ctx)
-    if(await startcheck(ctx, 'jv java')) return;
-  await jvStarter(bot, ctx) });
+jvScene.on("message", async (ctx: any) => {
+  cmdd(ctx)
+  if (await startcheck(ctx, 'jv java')) return;
+  await jvStarter(bot, ctx)
+});
 
 let jsScene = new Scenes.BaseScene<Scenes.SceneContext>("js");
-jsScene.enter(async (ctx: any) => { 
-   cmdd(ctx)
-    if(await startcheck(ctx, 'js node')) return;
-  await jsStarter(bot, ctx) });
+jsScene.enter(async (ctx: any) => {
+  cmdd(ctx)
+  if (await startcheck(ctx, 'js node')) return;
+  // return console.log(ctx.message)
+  await jsStarter(bot, ctx)
+});
 
-jsScene.on("message", async (ctx: any) => { 
-   cmdd(ctx)
-    if(await startcheck(ctx, 'js node')) return;
-  await jsStarter(bot, ctx) });
+jsScene.on("message", async (ctx: any) => {
+  cmdd(ctx)
+  if (await startcheck(ctx, 'js node')) return;
+  
+  await jsStarter(bot, ctx)
+});
 
 let tsScene = new Scenes.BaseScene<Scenes.SceneContext>("ts");
-tsScene.enter(async (ctx: any) => { 
-   cmdd(ctx)
-    if(await startcheck(ctx, 'ts type')) return;
-  await tsStarter(bot, ctx) });
+tsScene.enter(async (ctx: any) => {
+  cmdd(ctx)
+  if (await startcheck(ctx, 'ts type')) return;
+  await tsStarter(bot, ctx)
+});
 
-tsScene.on("message", async (ctx: any) => { 
-   cmdd(ctx)
-    if(await startcheck(ctx, 'ts type')) return;
-  await tsStarter(bot, ctx) });
+tsScene.on("message", async (ctx: any) => {
+  cmdd(ctx)
+  if (await startcheck(ctx, 'ts type')) return;
+  await tsStarter(bot, ctx)
+});
 
 let goScene = new Scenes.BaseScene<Scenes.SceneContext>("go");
-goScene.enter(async (ctx: any) => { 
-   cmdd(ctx)
-    if(await startcheck(ctx, 'go')) return;
-  await goStarter(bot, ctx) });
+goScene.enter(async (ctx: any) => {
+  cmdd(ctx)
+  if (await startcheck(ctx, 'go')) return;
+  await goStarter(bot, ctx)
+});
 
-goScene.on("message", async (ctx: any) => { 
-   cmdd(ctx)
-    await startcheck(ctx, 'go')
-  await goStarter(bot, ctx) });
+goScene.on("message", async (ctx: any) => {
+  cmdd(ctx)
+  await startcheck(ctx, 'go')
+  await goStarter(bot, ctx)
+});
 
 // making instance of Telegraf class
 let bot = new Telegraf<Scenes.SceneContext>(config.token);
@@ -133,43 +149,53 @@ bot.use(session());
 bot.use(stage.middleware());
 
 // Main Program starts from here it listens /js /py all commands and codes 
-bot.hears(new RegExp("^\\" + config.startSymbol + "(code|py|python|ts|type|js|node|cc|cpp|cplus|go|jv|java|c\\+\\+)", "i"), async (ctx: any) => {
+bot.hears(new RegExp("^\\" + config.startSymbol + "(code|start|py|python|ts|type|js|node|cc|cpp|cplus|go|jv|java|c\\+\\+)|\\/start", "i"), async (ctx: any, next: any) => {
   try {
-    
-  let compiler: any = ctx.message.text + "";
-  let memb = await ctx.getChatMember(ctx.botInfo.id)
-  if (!memb.can_delete_messages) {
-    if ((ctx.chat.id + "").startsWith("-100"))
-      return ctx.reply('I must be admin with delete message permission')
-  }
-    
-  function cmp(a:string){
-return (new RegExp("^\\" + config.startSymbol + a, "i")).test(compiler)}
-  
-  if (cmp("py|python"))
-    ctx.scene.enter("py")
-  else if (cmp("cc|code"))
-    ctx.scene.enter("code")
-  else if (cmp("js|node"))
-    ctx.scene.enter("js")
-  else if (cmp("ts|type")) {
-    ctx.reply("Excecuting typescript code..")
-      .then(async (m: any) => { await h.sleep(3000); return m.message_id })
-      .then((m: any) => { ctx.deleteMessage(m) })
-      .catch((err: any) => { })
+    let compiler: any = ctx.message.text + "";
 
-    await h.sleep(300)
-    ctx.scene.enter("ts")
-  }
-    
-  else if (cmp("cpp|cplus"))
-    ctx.scene.enter("cpp")
-  else if (cmp("jv|java"))
-    ctx.scene.enter("jv")
-  else if (cmp("go"))
-    ctx.scene.enter("go")
 
-      } catch (error) {
+    let memb = await ctx.getChatMember(ctx.botInfo.id)
+    if (!memb.can_delete_messages) {
+      if ((ctx.chat.id + "").startsWith("-100"))
+        return ctx.reply('I must be admin with delete message permission')
+    }
+
+    function cmp(a: string) {
+      return (new RegExp("^\\" + config.startSymbol + a, "i")).test(compiler)
+    }
+
+        console.log("yes")
+    if (compiler.startsWith("/start") && objj.hasOwnProperty("" + ctx.message.from.id)) {
+      let kkk = objj["" + ctx.message.from.id]
+      ctx.message.text = kkk.text;
+      return ctx.scene.enter(kkk.cmp)
+    }
+    console.log(objj)
+    if (cmp("py|python"))
+      ctx.scene.enter("py")
+    else if (cmp("cc|code"))
+      ctx.scene.enter("code")
+    else if (cmp("js|node"))
+      ctx.scene.enter("js")
+    else if (cmp("ts|type")) {
+      ctx.reply("Excecuting typescript code..")
+        .then(async (m: any) => { await h.sleep(3000); return m.message_id })
+        .then((m: any) => { ctx.deleteMessage(m) })
+        .catch((err: any) => { })
+
+      await h.sleep(300)
+      ctx.scene.enter("ts")
+    }
+
+    else if (cmp("cpp|cplus"))
+      ctx.scene.enter("cpp")
+    else if (cmp("jv|java"))
+      ctx.scene.enter("jv")
+    else if (cmp("go"))
+      ctx.scene.enter("go")
+    next();
+  } catch (error) {
+    console.log("error start index")
   }
 })
 
@@ -177,62 +203,63 @@ return (new RegExp("^\\" + config.startSymbol + a, "i")).test(compiler)}
 bot.launch({ dropPendingUpdates: true });
 
 // This function checks that if any compiler command change then it changes session; Example : js to py
-async function startcheck(ctx:any, y:any, json:any ={}) {
+async function startcheck(ctx: any, y: any, json: any = {}) {
   try {
-    
-if(!ctx.message.text.startsWith(config.startSymbol))
-  return false
-  
-  function cmp(a:string){
-return (new RegExp("^\\" + config.startSymbol + a, "i")).test(ctx.message.text)}
-  
-   let cmd = ctx.message.text.match(new RegExp("^\\"+ config.startSymbol + "[a-zA-Z0-9]{2,9}", 'i'))
-if(!cmd) return false;
-  
-  let cst = cmd[0].replace(config.startSymbol, "")
-   cst = cst.toLowerCase()
-  if(y.includes(cst)){
-    console.log('yes')
+
+    if (!ctx.message.text.startsWith(config.startSymbol))
+      return false
+
+    function cmp(a: string) {
+      return (new RegExp("^\\" + config.startSymbol + a, "i")).test(ctx.message.text)
+    }
+
+    let cmd = ctx.message.text.match(new RegExp("^\\" + config.startSymbol + "[a-zA-Z0-9]{2,9}", 'i'))
+    if (!cmd) return false;
+
+    let cst = cmd[0].replace(config.startSymbol, "")
+    cst = cst.toLowerCase()
+    if (y.includes(cst)) {
+      console.log('yes')
+      return false
+    }
+
+    if (("py python").includes(cst)) {
+      ctx.scene.enter("py")
+      return true
+    }
+    else if (("cc code").includes(cst)) {
+      ctx.scene.enter("code")
+      return true
+    }
+    else if (("js node").includes(cst)) {
+      ctx.scene.enter("js")
+      return true
+    }
+    else if (("ts type").includes(cst)) {
+      ctx.reply("Excecuting typescript code..")
+        .then(async (m: any) => { await h.sleep(3000); return m.message_id })
+        .then((m: any) => { ctx.deleteMessage(m) })
+        .catch((err: any) => { })
+
+      await h.sleep(300)
+      ctx.scene.enter("ts")
+      return true
+    }
+    else if (("cpp cplus").includes(cst)) {
+      ctx.scene.enter("cpp")
+      return true
+    }
+    else if (("jv java").includes(cst)) {
+      ctx.scene.enter("jv")
+      return true
+    }
+    else if (("go").includes(cst)) {
+      ctx.scene.enter("go")
+      return true
+    }
+
     return false
-  }
-  
-  if (("py python").includes(cst)){
-    ctx.scene.enter("py")
-    return true
-  }
-  else if (("cc code").includes(cst)){
-    ctx.scene.enter("code")
-    return true
-  }
-  else if (("js node").includes(cst)){
-    ctx.scene.enter("js")
-    return true
-  }
-  else if (("ts type").includes(cst)) {
-    ctx.reply("Excecuting typescript code..")
-      .then(async (m: any) => { await h.sleep(3000); return m.message_id })
-      .then((m: any) => { ctx.deleteMessage(m) })
-      .catch((err: any) => { })
-
-    await h.sleep(300)
-    ctx.scene.enter("ts")
-    return true
-  }
-  else if (("cpp cplus").includes(cst)){
-    ctx.scene.enter("cpp")
-    return true
-  }
-  else if (("jv java").includes(cst)){
-    ctx.scene.enter("jv")
-     return true
-  }
-  else if (("go").includes(cst)){
-    ctx.scene.enter("go")
-     return true
-  }
-
-  return false
-      } catch (error) {
+  } catch (error) {
     return false
   }
 }
